feat: separation of concern untuk part ujian
